requirements.txt
mongoengine
fastapi
uvicorn
requests
jsonpath
pydantic
python-jose
json5